+++ packages/core/src/errors/index.ts
@@ -1,3 +1,4 @@
 export { RouterError } from './router-error'
 export { NavigationFailure } from './navigation-failure'
 export { UniApiError, isUniApiError } from './uni-api-error'
+export { isNavigationFailure } from './is-navigation-failure'