+++ packages/uni-app/uni_modules/mxuni-router-v2/js_sdk/plugins/index.js
@@ -1,1 +1,1 @@
-export { AnimationPlugin, ChannelPlugin, InterceptorPlugin, ParamsPlugin, usePageChannel } from '../chunk-KMISK4GA.js';
+export { AnimationPlugin, ChannelPlugin, InterceptorPlugin, ParamsPlugin, usePageChannel } from '../chunk-OHFHI3IR.js';